fixed: remove extra files, separate quotes and images
