+++ backgrounds.js
@@ -1,0 +1,5 @@
+export const backgrounds = [
+    'url(images2/jason-blackeye-GPPAjJicemU-unsplash.jpg)',
+    'url(images2/cristian-giordano-qtqcSwzekFo-unsplash.jpg)',
+    'url(images2/lorenzo-lamonica-PJAXlV_H8-4-unsplash.jpg)'
+];
