+++ main.html
@@ -8,7 +8,7 @@
     <title>Just For Today</title>
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link rel="stylesheet" href="./main2.css">
+    <link rel="stylesheet" href="./main.css">
 </head>
 
 <body>
@@ -27,7 +27,7 @@
         <div id="quote" class="quoteAndSource"></div>
         <div id="quoteSource" class="quoteAndSource"></div>
     </div>
-    <script src="main2.js"></script>
+    <script src="main.js"></script>
 </body>
 
 </html>
