--- main1.css
@@ -1,41 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Rubik+Spray+Paint&display=swap');
-
-body {
-    background-color: #C9A992;
-    font-family: 'Rubik Spray Paint', cursive;
-}
-
-#theGrid {
-    display: grid;
-    grid-template-columns: 2fr 1fr 1fr 1fr 2fr;
-    grid-template-rows: 2fr 1fr 1fr 2fr;
-    grid-column-gap: 20px;
-}
-
-#clock {
-    grid-column: 2 / 5;
-    grid-row: 2;
-}
-
-#clock div {
-    display: inline-block;
-    color: #96C8EB;
-    font-size: 100px;
-
-}
-
-#clock span {
-    color: #FDF518;
-    font-size: 100px;
-}
-
-#units {
-    grid-column: 2 / 5;
-    grid-row: 3;
-    color: #EE5D88;
-    font-size: 30px;
-}
-
-#units div {
-    display: inline-block;
-}
