+++ quotes.js
@@ -1,0 +1,17 @@
+export const quotes = [
+    {text: `“Just believe in yourself. Even if you don’t, pretend that you do and, at some point, you will.”`,  source: `- Venus Williams`},
+    {text: `“Believe in yourself and all that you are. Know that there is something inside you that is greater than any obstacle.”`, source: `- Christian D. Larson`},
+    {text: `“Self-belief and hard work will always earn you success.”`, source: `- Virat Kohli`},
+    {text: `“If you think you can do a thing or think you can’t do a thing, you’re right.”` , source: `- Henry Ford`},
+    {text: `“Don’t let what you cannot do interfere with what you can do.”`, source: `- John Wooden`},
+    {text: `“Believe in yourself, take on your challenges, dig deep within yourself to conquer fears. Never let anyone bring you down. You got to keep going.”`, source: `- Chantal Sutherland`},
+    {text: `“Keep your dreams alive. Understand to achieve anything requires faith and belief in yourself, vision, hard work, determination, and dedication. Remember all things are possible for those who believe.”`, source: `- Gail Devers`},
+    {text: `“Successful and unsuccessful people do not vary greatly in their abilities. They vary in their desires to reach their potential.”`, source: `- John Maxwell`},
+    {text: `“If you don’t sacrifice for what you want, what you want becomes the sacrifice.”`, source: `- Unknown`},
+    {text: `“Life is never made unbearable by circumstance, but only by lack of meaning and purpose.”`, source: `- Viktor Frankl`},
+    {text:`“Forget all the reasons it won’t work and believe the one reason that it will.”`, source: `― Unknown`},
+    {text:`“Every morning you have two choices: continue to sleep with your dreams, or wake up and chase them.”`, source: `- Unkown`},
+    {text: `“Fear, inherently, is not meant to limit you. Fear is the brain’s way of saying that there is something important for you to overcome.”`, source: `- Unknown`},
+    {text: `“I don’t measure a man’s success by how high he climbs, but how high he bounces when he hits the bottom.”`, source: `- General George S. Patton`}
+]
+
